added height change, to reflow.
Resolve #46
